+++ src/components/CheckBox/CheckBox.js
@@ -1,32 +1,43 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
+import Icon from '../Icon';
+
+const getBackground = ({ disabled, checked, theme }) => {
+  if (disabled) return checked || checked === null ? theme.grey04 : theme.grey05;
+  return checked || checked === null ? theme.primary : theme.white;
+};
 
 const Wrapper = styled('span')`
   pointer-events: ${p => (p.disabled ? 'none' : 'initial')};
-  opacity: ${p => (p.disabled ? '0.5' : 'initial')};
   cursor: pointer;
   line-height: 1;
   display: inline-flex;
   align-items: center;
 `;
 
-const Input = styled('span')`
-  display: inline-block;
+const Input = styled(({ disabled, checked, ...otherProps }) => <span {...otherProps} />)`
+  display: inline-flex;
+  justify-content: center;
+  align-items: center;
   position: relative;
-  width: 0.8125em;
-  height: 0.8125em;
-  box-shadow: 0px 0px 0px 1px ${p => (p.checked ? p.theme.primary : p.theme.grey03)};
-  background: ${p => (p.checked ? p.theme.primary : 'transparent')};
+  width: 1em;
+  height: 1em;
+  border-radius: ${p => p.theme.borderRadius};
+  box-shadow: 0px 0px 0px 1px ${p => p.theme.grey04};
+  background: ${getBackground};
   border: solid 0.05em #fff;
   box-sizing: border-box;
   transition: background 0.3s ease-out;
+  color: ${p => (!p.checked ? 'transparent' : p.theme.white)};
+  fill: none;
 `;
 
-const Label = styled('span')`
+const Label = styled(({ disabled, ...otherProps }) => <span {...otherProps} />)`
   display: inline-block;
   margin-left: 8px;
   user-select: none;
+  opacity: ${p => (p.disabled ? '0.5' : 'initial')};
 `;
 
 const InputWrapper = styled('span')`
@@ -37,9 +48,11 @@ const InputWrapper = styled('span')`
 const CheckBox = ({ label, checked, disabled, onToggle, ...otherProps }) => (
   <Wrapper {...otherProps} onClick={onToggle} disabled={disabled}>
     <InputWrapper>
-      <Input checked={checked} />
+      <Input checked={checked} disabled={disabled}>
+        <Icon name="check" width="0.5em" height="0.5em" />
+      </Input>
     </InputWrapper>
-    <Label>{label}</Label>
+    <Label disabled={disabled}>{label}</Label>
   </Wrapper>
 );
 
@@ -49,7 +62,7 @@ CheckBox.defaultProps = {
 
 CheckBox.propTypes = {
   onToggle: PropTypes.func,
-  label: PropTypes.string,
+  label: PropTypes.node,
   disabled: PropTypes.bool,
   checked: PropTypes.bool,
 };